deprecate sdk::pubkey::{read_pubkey_file, write_pubkey_file} and duplicate in solana-keygen
